refactor(ui): split speed source sections
